requirements.txt
asgiref==3.8.1
Django==5.1.3
djangorestframework==3.15.2
psycopg==3.2.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
setuptools==75.6.0
sqlparse==0.5.2
tzdata==2024.2
